third_party/WebKit/LayoutTests/fast/canvas/draw-focus-if-needed-expected.txt

PASS canvas.drawFocusIfNeeded(null); did not throw exception.
PASS canvas.drawFocusIfNeeded(); threw exception TypeError: Failed to execute 'drawFocusIfNeeded' on 'CanvasRenderingContext2D': 1 argument required, but only 0 present..
PASS canvas.drawFocusIfNeeded(null); did not throw exception.
PASS data[0] != 0 || data[1] != 0 || data[2] != 0 is true
PASS data[0] == 0 && data[1] == 0 && data[2] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
